refactor: move lint module to utils
